+++ tests/test.py
@@ -94,18 +94,6 @@ document.write('<a href="' + href + '">' + linktext + "</a>");
         self.assertEqual(sp('"Isn\'t this fun?"'),
                          '&#8220;Isn&#8217;t this fun?&#8221;')
 
-    def test_deprecated_str_attr(self):
-
-        TEXT = '"foo" -- bar'
-
-        T = sp(TEXT, 'q')
-        E = '&#8220;foo&#8221; -- bar'
-        self.assertEquals(T, E)
-
-        T = sp(TEXT, 'qd')
-        E = '&#8220;foo&#8221; &#8212; bar'
-        self.assertEquals(T, E)
-
 
 def load_tests(loader, tests, pattern):